requirements.txt
sentence-transformers
numpy
google-generativeai
python-dotenv